refactor: separate app creation from actual server
