+++ src/server.ts
@@ -1,0 +1,8 @@
+import createApp from "./app.js";
+
+const PORT = process.env.PORT || 3000;
+const app = createApp();
+
+app.listen(PORT, () => {
+  console.log(`Server listening on port ${PORT}`);
+});
